ISSUE: var_lst is NULL in simple_command nodes
suspect that it is not initialise during parsing
