+++ srcs/execute/execute.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   execute.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zyeoh <zyeoh@student.42.fr>                +#+  +:+       +#+        */
+/*   By: sting <sting@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/11 16:34:17 by sting             #+#    #+#             */
-/*   Updated: 2024/05/29 17:13:24 by zyeoh            ###   ########.fr       */
+/*   Updated: 2024/05/30 10:39:39 by sting            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
