+++ srcs/execute/exit_status.c
@@ -6,7 +6,7 @@
 /*   By: sting <sting@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/09 14:46:03 by sting             #+#    #+#             */
-/*   Updated: 2024/05/17 15:22:33 by sting            ###   ########.fr       */
+/*   Updated: 2024/05/30 10:52:17 by sting            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,15 @@ void set_exit_status(int exit_code, t_var *var_lst)
 	char *updated;
 	t_var *exit_status;
 
+	print_var_lst(var_lst);
+	if (var_lst == NULL)
+		printf("var_lst is NULL in set_exit_status\n"); // ! ISSUE: var_lst is not assign for each individul node?
 	exit_status = get_var_node("?", var_lst);
+	if (exit_status == NULL)
+	{
+		ft_putstr_fd("Exit status node can't be found in var_lst\n", STDERR_FILENO);
+		return ;
+	}
 	str = ft_itoa(exit_code);
 	if_null_perror_n_exit(str, "ft_itoa", EXIT_FAILURE);
 	updated = (char *)malloc(1 + 1 + ft_strlen(str) + 1);
