+++ srcs/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zyeoh <zyeoh@student.42.fr>                +#+  +:+       +#+        */
+/*   By: sting <sting@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/15 16:11:50 by sting             #+#    #+#             */
-/*   Updated: 2024/05/29 18:31:00 by zyeoh            ###   ########.fr       */
+/*   Updated: 2024/05/30 10:51:07 by sting            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -94,7 +94,7 @@ int	main(int argc, char **argv, char **env)
 			break ;
 		shell_data.ast_root = create_ast(&shell_data);
 		print_ast(shell_data.ast_root, 0);
-		// execute(shell_data.ast_root);
+		execute(shell_data.ast_root);
 		// print_tokens(shell_data.token_root);
 		// free_tokens(shell_data.token_root);
 		shell_data.token_root = NULL;
