+++ srcs/utils/s_utils1.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils1.c                                           :+:      :+:    :+:   */
+/*   s_utils1.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sting <sting@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/22 18:33:19 by zyeoh             #+#    #+#             */
-/*   Updated: 2024/05/29 14:31:58 by sting            ###   ########.fr       */
+/*   Updated: 2024/05/30 10:33:50 by sting            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -80,17 +80,17 @@ int	arr_str_count(char **arr)
 	return (count);
 }
 
-int	print_env_var(t_var *var_lst, char *add_msg_before_var)
-{
-	while (var_lst != NULL)
-	{
-		if (var_lst->is_exported && ft_strchr(var_lst->str, '=') != NULL)
-		{
-			if (add_msg_before_var && add_msg_before_var[0] != '\0')
-				ft_printf("%s", add_msg_before_var);
-			ft_printf("%s\n", var_lst->str);
-		}
-		var_lst = var_lst->next;
-	}
-	return (EXIT_SUCCESS);
-}
+// int	print_env_var(t_var *var_lst, char *add_msg_before_var)
+// {
+// 	while (var_lst != NULL)
+// 	{
+// 		if (var_lst->is_exported && ft_strchr(var_lst->str, '=') != NULL)
+// 		{
+// 			if (add_msg_before_var && add_msg_before_var[0] != '\0')
+// 				ft_printf("%s", add_msg_before_var);
+// 			ft_printf("%s\n", var_lst->str);
+// 		}
+// 		var_lst = var_lst->next;
+// 	}
+// 	return (EXIT_SUCCESS);
+// }
